// src/hooks/orval/model/adminsRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */
import type { AdminsRecordRole } from './adminsRecordRole.ts';

export interface AdminsRecord {
  /** Unique record identifier */
  id?: string;
  /** @minLength 8 */
  password: string;
  /**
     * @minLength 30
     * @maxLength 60
     */
  tokenKey: string;
  email: string;
  emailVisibility?: boolean;
  verified?: boolean;
  name: string;
  role: AdminsRecordRole;
  is_active: boolean;
  last_login_at?: string;
  created?: string;
  updated?: string;
}